+++ Interfaces/Repository/IPaymentTermRepository.cs
@@ -5,7 +5,7 @@ namespace Interfaces.Repository
     /// <summary>
     /// Payment Term Repository Interface
     /// </summary>
-    public interface IPaymentTermRepository : IBaseRepository<PaymentTerm, int>
+    public interface IPaymentTermRepository : IBaseRepository<PaymentTerm, long>
     {
 
     }
